Split RPCTest into two and speed up test cases
