+++ src/python/grpcio_tests/tests/tests.json
@@ -65,7 +65,8 @@
   "unit._metadata_test.MetadataTest",
   "unit._reconnect_test.ReconnectTest",
   "unit._resource_exhausted_test.ResourceExhaustedTest",
-  "unit._rpc_test.RPCTest",
+  "unit._rpc_part_1_test.RPCPart1Test",
+  "unit._rpc_part_2_test.RPCPart2Test",
   "unit._server_shutdown_test.ServerShutdown",
   "unit._server_ssl_cert_config_test.ServerSSLCertConfigFetcherParamsChecks",
   "unit._server_ssl_cert_config_test.ServerSSLCertReloadTestCertConfigReuse",
