+++ src/python/grpcio_tests/tests/unit/BUILD.bazel
@@ -31,7 +31,8 @@ GRPCIO_TESTS_UNIT = [
     "_metadata_test.py",
     "_reconnect_test.py",
     "_resource_exhausted_test.py",
-    "_rpc_test.py",
+    "_rpc_part_1_test.py",
+    "_rpc_part_2_test.py",
     "_signal_handling_test.py",
     # TODO(ghostwriternr): To be added later.
     # "_server_ssl_cert_config_test.py",
@@ -75,6 +76,11 @@ py_library(
 )
 
 py_library(
+    name = "_rpc_test_helpers",
+    srcs = ["_rpc_test_helpers.py"],
+)
+
+py_library(
     name = "_server_shutdown_scenarios",
     srcs = ["_server_shutdown_scenarios.py"],
 )
@@ -96,6 +102,7 @@ py_library(
         main = test_file_name,
         deps = [
             ":_exit_scenarios",
+            ":_rpc_test_helpers",
             ":_from_grpc_import_star",
             ":_server_shutdown_scenarios",
             ":_signal_client",
